requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
torchsummary
sklearn
pyyaml
scikit-learn
Pillow
# tensorboard